arch_challenges/SqlInjection1/requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql-connector-repackaged==0.3.1
Werkzeug==0.11.4
wheel==0.29.0
